+++ contracts/library/Rarity.sol
@@ -5,19 +5,25 @@ import "../interfaces/core/IRarity.sol";
 import "./Attributes.sol";
 
 library Rarity {
-  IRarity constant RARITY = IRarity(0xce761D788DF608BD21bdd59d6f4B54b2e27F25Bb);
+    IRarity constant RARITY =
+        IRarity(0xce761D788DF608BD21bdd59d6f4B54b2e27F25Bb);
 
-  function level(uint summoner) public view returns (uint) {
-    return RARITY.level(summoner);
-  }
+    function level(uint256 summoner) public view returns (uint256) {
+        return RARITY.level(summoner);
+    }
 
-  function class(uint summoner) public view returns (uint) {
-    return RARITY.class(summoner);
-  }
+    function class(uint256 summoner) public view returns (uint256) {
+        return RARITY.class(summoner);
+    }
 
-  function isApprovedOrOwnerOfSummoner(uint summoner) public view returns (bool) {
-    if(RARITY.getApproved(summoner) == msg.sender) return true;
-    address owner = RARITY.ownerOf(summoner);
-    return owner == msg.sender || RARITY.isApprovedForAll(owner, msg.sender);
-  }
+    function isApprovedOrOwnerOfSummoner(uint256 summoner)
+        public
+        view
+        returns (bool)
+    {
+        if (RARITY.getApproved(summoner) == msg.sender) return true;
+        address owner = RARITY.ownerOf(summoner);
+        return
+            owner == msg.sender || RARITY.isApprovedForAll(owner, msg.sender);
+    }
 }
